# libs/core-gateway/src/main/kotlin/openBook/api/BookApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package openBook.api

import openBook.model.BookGetByInfoIdResponse
import openBook.model.BookGetByNameResponse
import openBook.model.BookInfo
import openBook.model.ErrorResponse
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired


import kotlinx.coroutines.flow.Flow
import kotlin.collections.List
import kotlin.collections.Map

@RequestMapping("\${api.base-path:}")
interface BookApi {


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/book/by-info/{book_info_id}"],
            produces = ["application/json"]
    )
    suspend fun v1BookByInfoBookInfoIdGet( @PathVariable("book_info_id") bookInfoId: java.util.UUID): ResponseEntity<BookGetByInfoIdResponse>


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/book/by-name/{book_name}"],
            produces = ["application/json"]
    )
    suspend fun v1BookByNameBookNameGet( @PathVariable("book_name") bookName: kotlin.String): ResponseEntity<BookGetByNameResponse>


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/book/info-details/{book_info_id}"],
            produces = ["application/json"]
    )
    suspend fun v1BookInfoDetailsBookInfoIdGet( @PathVariable("book_info_id") bookInfoId: java.util.UUID): ResponseEntity<BookInfo>
}